requirements.txt
asgiref==3.5.0
Django==4.0.4
django-braces==1.15.0
django-crispy-forms==1.14.0
django-tinymce==3.4.0
Pillow==9.1.0
psycopg2==2.9.3
sqlparse==0.4.2
tzdata==2022.1
